enable compiler warnings (#200)

+++ src/libserver/data/helper/ProtocolHelper.cpp
@@ -284,16 +284,17 @@ void BuildProtocolEgg(
   protocolEgg.uid = eggRecord.uid();
   protocolEgg.itemTid = eggRecord.itemTid();
 
-  protocolEgg.totalHatchingTime = std::chrono::duration_cast<std::chrono::seconds>(
-    hatchDuration).count();
+  protocolEgg.totalHatchingTime = static_cast<uint32_t>(std::chrono::duration_cast<std::chrono::seconds>(
+    hatchDuration).count());
 
   const auto totalHatchingDuration = std::chrono::system_clock::now() - eggRecord.incubatedAt();
   const auto totalBoostedDuration = eggRecord.boostsUsed() * std::chrono::hours(8);
   const auto hatchTimeRemaining = hatchDuration - totalHatchingDuration - totalBoostedDuration;
 
   protocolEgg.timeRemaining = std::max(
-    std::chrono::duration_cast<std::chrono::seconds>(hatchTimeRemaining).count(),
-    int64_t{0});
+    static_cast<uint32_t>(std::chrono::duration_cast<std::chrono::seconds>(
+      hatchTimeRemaining).count()),
+    uint32_t{0});
   
   protocolEgg.boost = 400000;
 }
@@ -309,9 +310,9 @@ void BuildProtocolSettings(
     for (const auto& keyboardBinding : settingsRecord.keyboardBindings().value())
     {
       auto& protocolBinding = settings.keyboardOptions.bindings.emplace_back();
-      protocolBinding.primaryKey = keyboardBinding.primaryKey;
-      protocolBinding.type = keyboardBinding.type;
-      protocolBinding.secondaryKey = keyboardBinding.secondaryKey;
+      protocolBinding.primaryKey = static_cast<uint8_t>(keyboardBinding.primaryKey);
+      protocolBinding.type = static_cast<uint8_t>(keyboardBinding.type);
+      protocolBinding.secondaryKey = static_cast<uint8_t>(keyboardBinding.secondaryKey);
       protocolBinding.unused = 0; // Unused
     }
   }
@@ -323,9 +324,9 @@ void BuildProtocolSettings(
     for (const auto& keyboardBinding : settingsRecord.gamepadBindings().value())
     {
       auto& protocolBinding = settings.gamepadOptions.bindings.emplace_back();
-      protocolBinding.primaryButton = keyboardBinding.primaryKey;
-      protocolBinding.type = keyboardBinding.type;
-      protocolBinding.secondaryButton = keyboardBinding.secondaryKey;
+      protocolBinding.primaryButton = static_cast<uint8_t>(keyboardBinding.primaryKey);
+      protocolBinding.type = static_cast<uint8_t>(keyboardBinding.type);
+      protocolBinding.secondaryButton = static_cast<uint8_t>(keyboardBinding.secondaryKey);
       protocolBinding.unused = 0; // Unused
     }
   }
